#135 Update log output format from environment variables
Introduced an update function that modifies Klogging configuration based on the environment variables which allow for runtime customization of log rendering and sinks. Also added unit tests to ensure the correctness of this functionality. While modifying the configuration, updated SinkConfiguration class to be a data class to support the copy method needed for immutability in Kotlin. Finally, adjusted senderFrom implementation in Sending class to make it more concise.

This change is made with the goal of making logging configurations more flexible and easily adjustable based on the runtime environment without requiring a code change.